+++ freemarker-generator-cli/src/main/java/org/apache/freemarker/generator/cli/config/TemplateDirectorySupplier.java
@@ -17,35 +17,28 @@
 package org.apache.freemarker.generator.cli.config;
 
 import org.apache.commons.io.FilenameUtils;
+import org.apache.freemarker.generator.base.FreeMarkerConstants.Configuration;
+import org.apache.freemarker.generator.base.FreeMarkerConstants.SystemProperties;
+import org.apache.freemarker.generator.base.util.StringUtils;
 
 import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import java.util.function.Supplier;
 
 import static java.util.Arrays.asList;
 import static java.util.stream.Collectors.toList;
+import static org.apache.freemarker.generator.base.util.StringUtils.isNotEmpty;
 
 /**
  * Determine a list of directories to load a relative template file in the following order
  * <ol>
  *    <li>User-defined template directory</li>
- *    <li>Current working directory</li>
  *    <li>~/.freemarker-generator</li>
  *    <li>Application installation directory</li>
  * </ol>
  */
 public class TemplateDirectorySupplier implements Supplier<List<File>> {
-
-    /** Installation directory of "freemarker-generator" when invoked with shell wrapper */
-    private static final String APP_HOME = "app.home";
-
-    /** Home directory of the user */
-    private static final String USER_HOME = "user.home";
-
-    /** The user's optional "freemarker-generator" directory */
-    private static final String USER_CONFIGURATION_DIR_NAME = ".freemarker-generator";
 
     /** Additional template directory, e.g. provided as command line parameter */
     private final String additionalTemplateDirName;
@@ -57,7 +50,7 @@ public class TemplateDirectorySupplier implements Supplier<List<File>> {
     @Override
     public List<File> get() {
         return templateLoaderDirectories().stream()
-                .filter(Objects::nonNull)
+                .filter(StringUtils::isNotEmpty)
                 .map(FilenameUtils::normalize)
                 .map(File::new)
                 .distinct()
@@ -74,31 +67,36 @@ public class TemplateDirectorySupplier implements Supplier<List<File>> {
     }
 
     private String additionalTemplatesDirectory() {
-        return additionalTemplateDirName != null ? new File(additionalTemplateDirName).getAbsolutePath() : null;
+        return isNotEmpty(additionalTemplateDirName) ? new File(additionalTemplateDirName).getAbsolutePath() : null;
     }
 
     private String userConfigDirectory() {
-        final String userHomeDir = System.getProperty(USER_HOME);
-        return new File(userHomeDir, USER_CONFIGURATION_DIR_NAME).getAbsolutePath();
+        final String userHomeDir = System.getProperty(SystemProperties.USER_HOME);
+        if (isNotEmpty(userHomeDir)) {
+            return new File(userHomeDir, Configuration.USER_CONFIGURATION_DIR_NAME).getAbsolutePath();
+        } else {
+            return null;
+        }
     }
 
     private String userConfigTemplatesDirectory() {
-        return userConfigDirectory() + "/templates";
-    }
-
-    private static String applicationDirectory() {
-        return System.getProperty(APP_HOME);
+        return templatesDirectory(userConfigDirectory());
     }
 
     private static String applicationTemplatesDirectory() {
-        return applicationDirectory() + "/templates";
+        final String appHomeDir = System.getProperty(SystemProperties.APP_HOME);
+        return templatesDirectory(appHomeDir);
+    }
+
+    private static String templatesDirectory(String baseDirName) {
+        if (isNotEmpty(baseDirName)) {
+            return new File(baseDirName, "templates").getAbsolutePath();
+        } else {
+            return null;
+        }
     }
 
     private static boolean isDirectory(File directory) {
-        boolean exists = directory.exists();
-        boolean isDirectory = directory.isDirectory();
-        boolean canRead = directory.canRead();
-        boolean r = directory != null && exists && isDirectory && canRead;
-        return r;
+        return directory != null && directory.exists() && directory.isDirectory() && directory.canRead();
     }
 }